fixing save functionality
