+++ x-pack/plugins/apm/server/lib/settings/custom_link/create_or_update_custom_link.ts
@@ -4,12 +4,8 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import {
-  SERVICE_NAME,
-  SERVICE_ENVIRONMENT,
-  TRANSACTION_NAME,
-  TRANSACTION_TYPE
-} from '../../../../common/elasticsearch_fieldnames';
+import { pick } from 'lodash';
+import { filterOptions } from '../../../routes/settings/custom_link';
 import { APMIndexDocumentParams } from '../../helpers/es_client';
 import { Setup } from '../../helpers/setup_request';
 import { CustomLink } from './custom_link_types';
@@ -32,10 +28,7 @@ export async function createOrUpdateCustomLink({
       '@timestamp': Date.now(),
       label: customLink.label,
       url: customLink.url,
-      [SERVICE_NAME]: customLink[SERVICE_NAME],
-      [SERVICE_ENVIRONMENT]: customLink[SERVICE_ENVIRONMENT],
-      [TRANSACTION_NAME]: customLink[TRANSACTION_NAME],
-      [TRANSACTION_TYPE]: customLink[TRANSACTION_TYPE]
+      ...pick(customLink, filterOptions)
     }
   };
