+++ x-pack/plugins/apm/server/lib/settings/custom_link/custom_link_types.d.ts
@@ -4,9 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 import * as t from 'io-ts';
-import { FilterOptions } from '../../../routes/settings/custom_link';
-
-export type FilterOptionsType = t.TypeOf<typeof FilterOptions>;
+import { FilterOptionsType } from '../../../routes/settings/custom_link';
 
 export type CustomLink = {
   id?: string;
